+++ kernel/src/io/namesrv.rs
@@ -4,10 +4,10 @@
 
 use crate::ipc::{port, Message};
 use super::protocol::*;
-use core::sync::atomic::{AtomicU32, Ordering};
+use core::sync::atomic::{AtomicU64, Ordering};
 
 /// Global port ID for the name server.
-pub static NAMESRV_PORT: AtomicU32 = AtomicU32::new(u32::MAX);
+pub static NAMESRV_PORT: AtomicU64 = AtomicU64::new(u64::MAX);
 
 const MAX_SERVICES: usize = 32;
 const MAX_SVC_NAME: usize = 24;
@@ -15,7 +15,7 @@ const MAX_SVC_NAME: usize = 24;
 struct ServiceEntry {
     name: [u8; MAX_SVC_NAME],
     name_len: usize,
-    port_id: u32,
+    port_id: u64,
     active: bool,
 }
 
@@ -43,7 +43,7 @@ impl NameTable {
         }
     }
 
-    fn register(&mut self, name: &[u8], port_id: u32) -> bool {
+    fn register(&mut self, name: &[u8], port_id: u64) -> bool {
         // Check for duplicate — update if exists.
         for i in 0..self.count {
             if self.entries[i].active && self.entries[i].name_len == name.len()
@@ -66,7 +66,7 @@ impl NameTable {
         true
     }
 
-    fn lookup(&self, name: &[u8]) -> Option<u32> {
+    fn lookup(&self, name: &[u8]) -> Option<u64> {
         for i in 0..self.count {
             if self.entries[i].active && self.entries[i].name_len == name.len()
                 && &self.entries[i].name[..name.len()] == name
@@ -95,8 +95,8 @@ pub fn namesrv_server() -> ! {
         match msg.tag {
             NS_REGISTER => {
                 let name_len = (msg.data[3] & 0xFFFF_FFFF) as usize;
-                let reply_port = (msg.data[3] >> 32) as u32;
-                let service_port = msg.data[2] as u32;
+                let reply_port = msg.data[3] >> 32;
+                let service_port = msg.data[2];
                 let name_buf = unpack_name(msg.data[0], msg.data[1], 0, name_len);
                 let name = &name_buf[..name_len.min(MAX_SVC_NAME)];
 
@@ -106,21 +106,21 @@ pub fn namesrv_server() -> ! {
 
             NS_LOOKUP => {
                 let name_len = (msg.data[3] & 0xFFFF_FFFF) as usize;
-                let reply_port = (msg.data[3] >> 32) as u32;
+                let reply_port = msg.data[3] >> 32;
                 let name_buf = unpack_name(msg.data[0], msg.data[1], msg.data[2], name_len);
                 let name = &name_buf[..name_len.min(MAX_SVC_NAME)];
 
-                let port_id = table.lookup(name).unwrap_or(u32::MAX);
+                let port_id = table.lookup(name).unwrap_or(u64::MAX);
 
                 // Grant SEND cap for the looked-up service port to the client task.
-                if port_id != u32::MAX {
+                if port_id != u64::MAX {
                     if let Some(client_task) = port::port_creator(reply_port) {
                         let mut caps = crate::cap::CAP_SYSTEM.lock();
                         caps.grant_send_cap(client_task, port_id);
                     }
                 }
 
-                let _ = port::send_nb(reply_port, Message::new(NS_LOOKUP_OK, [port_id as u64, 0, 0, 0, 0, 0]));
+                let _ = port::send_nb(reply_port, Message::new(NS_LOOKUP_OK, [port_id, 0, 0, 0, 0, 0]));
             }
 
             _ => {}
